Add Web3Provider reat component
