+++ packages/react/src/hooks/useWallets.ts
@@ -1,0 +1,7 @@
+import type { WalletState } from '@web3-onboard/core'
+
+import { useAppState } from './useAppState'
+
+export const useWallets = (): WalletState[] => {
+  return useAppState('wallets')
+}
